requirements.txt
alembic==1.5.5
attrs==20.3.0
click==7.1.2
Flask==1.1.2
Flask-JWT==0.3.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
idna==3.1
importlib-metadata==3.7.2
itsdangerous==1.1.0
Jinja2==2.11.3
jsonpointer==2.0
jsonschema==3.2.0
Mako==1.1.4
MarkupSafe==1.1.1
mysqlclient==2.0.3
PyJWT==1.4.2
pyrsistent==0.17.3
python-dateutil==2.8.1
python-editor==1.0.4
rfc3987==1.3.8
six==1.15.0
SQLAlchemy==1.3.23
strict-rfc3339==0.7
typing-extensions==3.7.4.3
webcolors==1.11.1
Werkzeug==1.0.1
zipp==3.4.1
